requirements.txt
aiogram==2.25.1
pyexiftool==0.5.3
python-dotenv==1.0.1
